L13 : More senses for AI (Reacting to Damage)
SAttributeComponent::ApplyHealthChange now also includes an InstigatorActor parameter for informing who was that damaged/healed the owner of the component.
Updated HealthPotion and MagicProjectile C++ classes to use the new parameter InstigatorActor when calling ApplyHealthChange on an AttributeComponent.
Updated SAICharacter to target the last actor that damaged it. (Note : currently the AI is always targeting the PlayerCharacter since the QueryContextTargetActor asset is by default returning the player pawn as TargetActor. However this would indeed work as intended in case of a multiplayer game.)