L14 : 3D (Projected) Health bar with C++
Added a simple UMG health bar that appears on an enemy whey they get hit:

SWorldUserWidget class contains the projection code for the widget.
MinionHealth_Widget inherits from SWorldUserWidget and contains a sizebox and the logic for updating the material parameter for the health bar similar to how it was done in PlayerHealth_Widget for PlayerCharacter.
Updated SAICharacter to show the HealthBar widget when hit.