L15 : Refining Player Respawns
Updated SGameModeBase to include a method that give the player a new pawn.
The new GameMode functionality is triggered by SAttributeComponent.

Note : a slight bug is introduced due to this change as multiple Main_HUD widgets get created and stacked on top of each other (one for each new player pawn created). Will be fixed in a future commit.