L16 : Starting our Action System and first "Action" (Sprinting)
Created base class SAction (derived from UObject) to be used as base for all Actions/Abilities/Attacks. An Action be Started, Stopped and has an FName as an identifier.
Created SActionComponent that holds a TArray of available actions. The ActionComponent can be used to Start/Stop an action by name and/or add a new Action in the TArray of available actions.
Added an SActionComponent to SCharacter and invoke Sprint Action Start/Stop via it.
Created child BP object class Action_Sprint (responsible for sprinting) and granted that on BeginPlay in PlayerCharacter BP (via ActionComponent AddAction)
Created new key binding for Spring action in DefaultInput.ini .