L18 : Creating Buffs & Debuffs (Damage over Time Effect)
Created SActionEffect class for Applying Buffs/Debuffs over a specific period of time.
Updated SMagicProjectile class to apply an action effect (e.g. burning) when hitting a target.
Updated SActionComponent with function to RemoveAction as well as check if an Action shall be autostarted when added.
Created Effect_Burning BP inherited from SActionEffect and assigned that to Proj_MagicProjectile BP.
Created Status.Burning GameplayTag.

Note : The ActionComponent added previously to SAICharacter is now responsible for the burning effect when AI is hit by MagicProjectile.
The Effect_Burning (which inherits from SActionEffect) is added to the AI's ActionComponent (on MagicProjectile explosion) and then deletes itself from the Actions array of the ActionComponent when its effect timer has elapsed.