L20 : Authority and Actor instances on Clients
SetReplicates(true) for SMagicProjectile and SPowerupActor.
This spawns the powerups on both client and server (previously it was spawning only on the server since the spawning happens in GameMode which only exists on the server).
Also spawning of SMagicProjectile from the server is now visible on the client (client spawning is still not working as it needs a server RPC).
Updated PlayerControllerBP to only create HUD on each client via isLocalController check. This fixes a MessageLog error when exiting play.