L22 :  Limiting Authority of the Client
Updated SAction to also pass Instigator to other clients when starting an action (so attack anim is synced correctly as it needs the Instigator).
To achieve this the FActionRepData has been created to replicate both the bIsRunning state as well as the Instigator in one variable.
Updated OnRep_IsRunning->OnRep_RepData() and other uses of bIsRunning respectively.
Reduced debug logging in Start/Stop Action implementation functions.

Updated SActionComponent to simplify its debug log messages.

Updated SAction_ProjectileAttack to only spawn the projectile actor on server , while still showing the cosmetic stuff (anim + particles) on each client.

Updated SAttributeComponent to apply a HealthChange on server side only. This ensures a client cannot modify their own health (e.g. lose health because of a miscalculation on client side or e.g. cheating to boost their health).

Updated PlayerCharacter BP default actions to Action_MagicProjectile from Action_MultiplayerTest.

NOTE : There seems to be a bug in UHT where simply mentioning "GetLifetimeReplicatedProps" (even if just in a comment) would result in UHT not generating the necessary function declaration in generated.h which would in turn result in a compilation error. Maybe investigate further on this and report back to Epic with a fix.