L23 : Serializing World Transforms (World State #1)
Updated SSaveGame with a field to store actors to disk with their current transform at the time of saving.
A new struct (FActorSaveData) was created for that purpose.

Updated SGameModeBase to save/load actors (their transforms) in world that implement USGameplayInterface.

Updated TreasureChest BP (BaseMesh) to simulate physics. TreasureChest inherits from SItemChest which implements USGameplayInterface. That makes it ideal for testing the saving/loading of actor transforms.