L25 : Finishing Buff Bar & Animated materials for UI
Updated SAction and SActionEffect to track when (time) an action has started and remaining duration of effect respectively.

Used this new accessible info in EffectSlot widget to properly show the duration of a status effect in UI.
EffectSlot widget is now responsible for listening for StopAction for the effect it represents in the UI and remove itself from viewport when that time comes.

Updated EffectContainer to avoid warnings in MessageLog about default EffectSlots being present (as noted in previous commit message).

Assigned icon to Effect_Burning bp action.
Created M_BuffIcon material to have an animation to status icons in the UI showcasing their remaining durations. M_BuffIcon gets used in EffectSlot_Widget.

Slight UI adjustment in MinionHealth_Widget.
Added a comment in EffectContainer_Widget.