L28 : Cooking Content & Packaging
Updated "Monsters" PrimaryAssetType to AlwaysCook content when packaging.

Added  `!PrimaryAssetTypesToScan=ClearArray` at [/Script/Engine.AssetManagerSettings] in DefaultGame.ini to solve packaging issue with overlapping PrimaryAssetType for "Map" and "PrimaryAssetLabel". This (engine) issue seems to have been resolved in 4.27+ so this hack should not be necessary from that version onwards.

Tried a packaged build to ensure everything goes smoothly (it does :) ).