L29 : Improving the SaveGame, Collisions & UI
Updated Collision Profiles. Added new Projectile Collision Object Channel/Type which overlaps by default (DefaultEngine.ini).

Now SCharacter and SAICharacter both will react to overlap events of projectiles on their mesh and won't use the default WorldDynamic channel anymore.

Updated SGameModeBase and MainMenu_Widget to allow the option of loading a specific save game slot (currently no UI exists for selecting a specific slot but the backend code in C++ is ready for it).

Updated DamagePopup_Widget and SWorldUserWidget to fix an issue where the damage popup would appear in the top left corner of the screen (position : 0,0) when the enemy/damage popup was not rendered on the screen. The (BP) fix in DamagePopup_Widget was needed as DamagePopup_Widget was inheriting directly from UserWidget and not from SWorldUserWidget.